UI: only render input if control is in focus.
Change UI so that a control panel only has an input box when it is in focus. Otherwise, a panel displaying the current value of that parameter is rendered.

Factored out parts of the makePanel (was makeControl) function for readability.
Deleted the results section, since that's no longer needed!

Rather than having the main three parameters be Maybes, changing to Justs when the user inputs something, they're now just floats, with valid defaults set.