6: Allow only numbers to be entered into inputs
This prevents the user from entering anything but a number, BUT it does not preclude them entering a negative number.
The min attribute on the input only controls the numbers that can be reached when using the stepper arrow - the user can still manually enter a negative number.
The absolute value of the entered number is set on the model. Therefore, if a user enters -2 feet, calculations proceed as if 2 feet was entered.

The step is also set individually for each input type (footage, duration, framecount).