6: Use a number input in order to auto-validate input values
I've made the HTML input of type number, meaning that only numbers can be entered.

Even with the min attribute set, though, it doesn't preclude the user entering a negative number. So in the update function, I'm taking the absolute value of the entered number and using that for calculations and on the model.
Therefore, if a user enters -4, it's automatically converted to 4.

I've added CSS to suppress the stepper arrows that appear in number inputs in some browsers (including Chrome).