4: Use a text input for the duration control
Although anything can be typed in, only a string in the format hh:mm:ss will be accepted by the logic in the update function. Therefore, if something invalid is typed in, there won't be a visual indication that the input is wrong, but the model won't be updated.

This has less inbuilt validation than the time input approach, and the same copy/paste problem, but it looks nicer IMO.

Currently, if a number greater than 59 is entered for minutes or seconds, the whole value is rejected. It should probably be rounded to the largest possible value, i.e., 59.