4: Make param values clickable without changing control in focus
Previously, when a footage/duration/framecount panel was an info section, rather than an input , it was hard to select the parameter value - as soon as you clicked there, the panel became in focus and therefore became an input panel.

Now, I've stopped propagation on the div which displays the parameter value, so you can click there and copy the text without it flipping to an input section. If you click anywhere else on the panel, the onClick ChangeControlInFocus message is delivered as usual.