5: Enable Bulma navbar
The menu dropdown is open/closed by toggling the CSS class "is-active". In order to toggle this in the HTML, I added a Boolean flag on the model that is toggled when one clicks on the burger icon.
This is in order to keep everything in Elm rather than also editing the JavaScript.

I added styling to align the menu items right and narrow the width of the dropdown.
Thanks to the good folks of Stack Overflow: https://stackoverflow.com/questions/48735750/bulma-navbar-how-to-style-the-menu-on-dropdown

Also:
- Bump Bulma from 0.5.3 → 0.6.1
- Fix typo: métres → mètres