2 (bugfix): properly implement metric/imperial conversion
Previously, toggling metric/imperial via the buttons would change the SystemOfMeasurement stored on the model, but no recalculation was actually being performed. In the view, a conversion function changed things cosmetically, but the actual footage value wasn't changing.
Now, when the system is changed, the footage is recalculated. Therefore, if model.system is Imperial, model.footage always gives the length in feet; if model.system is Metric, model.footage is always in metres.

TODO: write some tests for calculating on metres rather than feet.