14: Remove need to click to focus; just have input boxes
Previously, each film parameter panel (footage, duration, frame count) needed to be clicked on to be 'focused', at which time it was coloured green and the input box was visible. The other two (non-active) panels just displayed values - no input box.

Obviously, it's nice to avoid the need to click twice. So now, I've removed the idea of a panel being in focus. Each of the three panels has an input boxe visible all the time, and when you change the value in one of them, the others update too.
This also had the advantage of simplifying the view code a LOT.

For the time being, the duration parameter is only available as seconds rather than hh:mm:ss. I'll enable the hh:mm:ss format again in future.

Addresses https://github.com/kfrn/film-figures/issues/14